% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{slide_dim}
\alias{slide_dim}
\title{Epiverse-TRACE default plotting dimensions for slides}
\format{
\subsection{\code{slide_dim}}{

A list with 2 elements:
\describe{
\item{slide_height}{A single \code{numeric} for the height of slide figures.}
\item{slide_width}{A single \code{numeric} for the width of slide figures.}
}
}
}
\usage{
slide_dim
}
\description{
Epiverse-TRACE default plotting dimensions for slides
}
\keyword{datasets}
