% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R, R/tracetheme-package.R
\docType{package}
\name{tracetheme-package}
\alias{tracetheme-package}
\title{tracetheme: A 'ggplot2' Theme and Scale Using 'Epiverse-TRACE' Visual Style}
\description{
A 'ggplot2' theme and scale using 'Epiverse-TRACE' visual style, colour scheme, and visualisation guidelines.

A 'ggplot2' theme and scale using 'Epiverse-TRACE' visual style, colour scheme, and visualisation guidelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/tracetheme}
  \item Report bugs at \url{https://github.com/epiverse-trace/tracetheme/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/tracetheme}
  \item Report bugs at \url{https://github.com/epiverse-trace/tracetheme/issues}
}

}
\author{
\strong{Maintainer}: Hugo Gruson \email{hugo@data.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID})

Authors:
\itemize{
  \item The Epiverse-TRACE team [copyright holder]
}

}
\keyword{internal}
