% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_dim}
\alias{vignette_dim}
\title{Epiverse-TRACE default plotting dimensions for vignettes}
\format{
\subsection{\code{vignette_dim}}{

A list with 2 elements:
\describe{
\item{vig_height}{A single \code{numeric} for the height of vignette figures.}
\item{vig_width}{A single \code{numeric} for the width of vignette figures.}
}
}
}
\usage{
vignette_dim
}
\description{
Epiverse-TRACE default plotting dimensions for vignettes
}
\keyword{datasets}
