% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tracetheme-package}
\alias{tracetheme}
\alias{tracetheme-package}
\title{tracetheme: A 'ggplot2' Theme and Scale Using 'Epiverse-TRACE' Visual Style}
\description{
A 'ggplot2' theme and scale using 'Epiverse-TRACE' visual style, colour scheme, and visualisation guidelines.
}
\author{
\strong{Maintainer}: The Epiverse-TRACE team \email{hugo@data.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID})

}
\keyword{internal}
